Moved the members import dialog out of its custom-fields subdirectory
ref https://linear.app/ghost/issue/BER-3897

The dialog lived in a custom-fields subdirectory because it was one of two import
implementations and that was what told it apart from the other. With the other one gone the
name described nothing, so the dialog moves up to sit as the only import dialog. The module
it kept alongside it, which added to the shared mapping vocabulary, went with it: a
re-export layer between one caller and one module is just indirection, and its test was a
duplicate of the shared one apart from the few cases that moved across. The comments still
describing two implementations were corrected at the same time.
